ai-hub-institute/pyproject.toml
[project]
name = "ai-hub-institute"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents[litellm]>=0.0.13",
]
